removed whitespace

+++ src/components/clientsComponents/ClientEdit.js
@@ -3,19 +3,19 @@ import { connect } from "react-redux"
 import { updateClient, deleteClient } from "../../action/clientAction"
 import { Link } from "react-router-dom"
 
-const ClientEdit = ({ updateClient, deleteClient, punches, drivers, history, match, clients, isLoggedIn }) => {
+const ClientEdit = ({ updateClient, deleteClient, punches, drivers, history, match, clients, isLoggedIn, current_user }) => {
     const [fullName, setfullName] = useState()
     const [payRate, setPayRate] = useState()
     const [driverId, setDriver] = useState()
     const client = clients.find(client => client.id === match)
     let displayClient = () => {
-        if (client !== undefined){
+        if (client !== undefined) {
             return (
                 <>
                     <h1>Edit {client.full_name}</h1>
                     <form onSubmit={e => handleSubmit(e)}>
                         <select name="driver_id" onChange={e => setDriver(e.currentTarget.selectedOptions[0].id.split("key")[1])}>
-                        <option default >Select Driver</option>
+                            <option default >Select Driver</option>
                             {listOfDrivers()}</select> <br />
                         <input name="full_name" placeholder="fullname" onChange={e => setfullName(e.target.value)} /> <br />
                         <input type="number" name="pay_rate" placeholder="pay_rate" onChange={e => setPayRate(e.target.value)} /> <br />
@@ -24,15 +24,16 @@ const ClientEdit = ({ updateClient, deleteClient, punches, drivers, history, mat
                     </form>
                     <br />
                     <br />
-                    
-                        <select name="id" onChange={e => createPunchForm(e)}>
+
+                    <select name="id" onChange={e => createPunchForm(e)}>
                         <option default >Select Punch</option>
-                            {listOfPunches()}</select>
-        
+                        {listOfPunches()}</select>
+
                 </>
-            )} else {
-                return <div className="loading"></div>
-            }
+            )
+        } else {
+            return <div className="loading"></div>
+        }
     }
     function listOfDrivers() {
         return drivers.map(driver => {
@@ -41,9 +42,26 @@ const ClientEdit = ({ updateClient, deleteClient, punches, drivers, history, mat
             )
         })
     }
-    
-    function createPunchForm(e){
-        // find the punch to edit
+    function displayLocalTime(mins) {
+        if (mins) {
+            let localTime = 0
+            let hours = Number(parseInt(mins.split("").slice(0, 2).join("")))
+            let minutes = Number(parseInt(mins.split("").slice(2, 4).join("")))
+            if (minutes < 10) minutes = (`0${minutes}`)
+
+            if (hours >= 12 && hours < 24) {
+                if (hours > 12) hours = hours - 12
+                localTime = `${hours}:${minutes} PM`
+            }
+            else if (hours < 13 || hours === 24) {
+                if (hours === 24) hours = hours - 12
+                localTime = `${hours}:${minutes} AM`
+            }
+            return localTime
+        }
+    }
+
+    function createPunchForm(e) {
         let findThePunches = punches.filter(punch => punch.client_id === client.id)
         let thisPunch = findThePunches.find(punch => punch.id ===
             Number(parseInt(e.currentTarget.selectedOptions[0].id.split("key")[1])))
@@ -54,12 +72,12 @@ const ClientEdit = ({ updateClient, deleteClient, punches, drivers, history, mat
 
     function listOfPunches() {
         let findThePunches = punches.filter(punch => punch.client_id === client.id)
-        return findThePunches.map(punch=> {
-            return(
-            
-                 <option key={punch.id} id={"key"+punch.id}> 
-                     id: {punch.id} Clocked In: {punch.clock_in} Clocked Out: {punch.clock_out}
-                 </option>
+        return findThePunches.map(punch => {
+            return (
+
+                <option key={punch.id} id={"key" + punch.id}>
+                    Clocked In: {displayLocalTime(punch.clock_in)} Clocked Out: {displayLocalTime(punch.clock_out)}
+                </option>
             )
         })
     }
@@ -72,8 +90,10 @@ const ClientEdit = ({ updateClient, deleteClient, punches, drivers, history, mat
             pay_rate: payRate,
             driver_id: driverId
         }
-        updateClient(updatedClient)
-        setTimeout(() => history.push("/clients/" + updatedClient.id), 2300)
+        if (current_user === client.driver.user_id) {
+            updateClient(updatedClient)
+            setTimeout(() => history.push("/clients/" + updatedClient.id), 2300)
+        }
     }
 
     function handleDelete(e) {
@@ -84,8 +104,10 @@ const ClientEdit = ({ updateClient, deleteClient, punches, drivers, history, mat
             pay_rate: Number(parseInt(payRate)),
             driver_id: Number(parseInt(driverId))
         }
-        deleteClient(deletingClient)
-        setTimeout(() => history.push("/clients"), 2300)
+        if (current_user === client.driver.user_id) {
+            deleteClient(deletingClient)
+            setTimeout(() => history.push("/clients"), 2300)
+        }
     }
 
     let redirect = () => {
